modified:   Aptech/JAVA/Example_java/.idea/workspace.xml 	modified:   Aptech/JAVA/Example_java/exceptionhandeling/Nest.java 	modified:   Aptech/JAVA/Example_java/Array And String/Boxing4.java 	modified:   Aptech/JAVA/Example_java/exceptionhandeling/TestMultipleCatchBlock.java 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/A.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/A1.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/A23.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Account.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Animal.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/ArithmeticException.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/BankDemo.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Boxing4.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/BoxingExample1.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Calculation12.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/CheckingAccount.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/ClassCastException.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Dog.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Employee.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Even_Odd_Number/even_detector.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Even_Odd_Number/main.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Even_Odd_Number/odd_detector.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Even_Odd_Number/option_choose.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/ExcepTest.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/ExcepTest1.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/GFG.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/InsufficientFundsException.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/IntArray.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Rectangle.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Rectangle2.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringBuilderExample.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringBuilderExample2.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringBuilderExample3.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringBuilderExample4.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringBuilderExample5.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringBuilderExample6.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringBuilderExample7.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringDemo.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringDemo1.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringDemo2.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringDemo3.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringFormatDemo.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/StringMethodsDemo.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Student.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Student1.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Student2.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Student3.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Student4.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Student5.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestAccount.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestArray.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestArray1.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestCollection1.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestCollection2.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestCollection3.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestCollection4.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestCollection5.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestCollection6.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestEmployee.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestOneDArray.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestRectangle1.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestRectangle2.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestStudent1.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestStudent2.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestStudent3.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestStudent4.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestThis2.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestThis4.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestThis5.class     	new file:   Aptech/JAVA/Example_java/out/production/Example_java/TestThis6.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/ThisStudent.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/UnboxingExample1.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/WrapperExample1.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/Unboxing.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/calculator/first.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/UnboxingExample2.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/calculator/option.class 	new file:   Aptech/JAVA/Example_java/out/production/Example_java/WrapperExample2.class
	new file:   Aptech/JAVA/Example_java/out/production/Example_java/calculator/main.class
	new file:   Aptech/JAVA/Example_java/out/production/Example_java/my_Account.class

+++ Aptech/JAVA/Example_java/Array And String/Boxing4.java
@@ -5,7 +5,8 @@ class Boxing4
   
   public static void main(String args[])
   {  
-      int a=30;  
-      m(a);  
+      //int a=30;
+      long a=30;
+      m(a);
   }
 }
